Move dot and cross into core/math.py
