+++ src/scipp/core/__init__.py
@@ -77,8 +77,8 @@ from .counts import counts_to_density, density_to_counts
 from .cumulative import cumsum
 from .dataset import irreducible_mask, merge
 from .groupby import groupby
-from .math import abs, nan_to_num, norm, reciprocal, pow, sqrt, exp, log, log10, round, floor, ceil, erf, erfc, midpoints
-from .operations import dot, islinspace, issorted, allsorted, cross, sort, values, variances, stddevs, rebin, where, to
+from .math import abs, cross, dot, nan_to_num, norm, reciprocal, pow, sqrt, exp, log, log10, round, floor, ceil, erf, erfc, midpoints
+from .operations import islinspace, issorted, allsorted, sort, values, variances, stddevs, rebin, where, to
 from .reduction import mean, nanmean, sum, nansum, min, max, nanmin, nanmax, all, any
 from .shape import broadcast, concat, fold, flatten, squeeze, transpose
 from .trigonometry import sin, cos, tan, asin, acos, atan, atan2
