+++ src/scipp/core/math.py
@@ -21,6 +21,28 @@ def abs(x: VariableLike, *, out: Optional[VariableLike] = None) -> VariableLike:
     return _call_cpp_func(_cpp.abs, x, out=out)
 
 
+def cross(x: VariableLike, y: VariableLike) -> VariableLike:
+    """Element-wise cross product.
+
+    :param x: Left hand side operand.
+    :param y: Right hand side operand.
+    :raises: If the dtype of the input is not vector3.
+    :return: The cross product of the input vectors.
+    """
+    return _call_cpp_func(_cpp.cross, x, y)
+
+
+def dot(x: VariableLike, y: VariableLike) -> VariableLike:
+    """Element-wise dot product.
+
+    :param x: Left hand side operand.
+    :param y: Right hand side operand.
+    :raises: If the dtype of the input is not vector3.
+    :return: The dot product of the input vectors.
+    """
+    return _call_cpp_func(_cpp.dot, x, y)
+
+
 def nan_to_num(x: _cpp.Variable,
                *,
                nan: _cpp.Variable = None,
